move plex webhook, and logger interface
